avance de ayuda en pase

+++ server/headers/model/Ball.h
@@ -43,8 +43,8 @@ class Ball {
         bool isInterceptable();
         void restart(int position);
         void isIntercepted(Player* player);
-        void isPassed(int direction, int passPower, bool highPass);
-        void isPassed(Velocity* velocity, int passPower, bool highPass);
+        void isPassed(int direction, float passPower, bool highPass);
+        void isPassed(Velocity* velocity, float passPower, bool highPass);
         void updatePosition();
 
         void calculateDominatedPosition();
